Add files via upload
Added AngleOfParallelism class in PD_HP.py, and skeleton code to include canvas functionality. Added area_of_h_polygon in PD_HP.py and edited PDBCConvexHull in MainFile.py to make it print the area of the drawn polygon. Need to update it to work with dragableDots and to setlabel with area, instead of printing it to console.
Added debugVariable to test canvas functionality of AngleOfParallelism, but still didn't manage to make it work yet.